+++ baybe/diagnostics/shap.py
@@ -1,4 +1,4 @@
-"""Diagnostics utilities."""
+"""SHAP utilities."""
 
 import numbers
 import warnings
@@ -32,6 +32,9 @@ def explainer(
 
     Raises:
         ValueError: If no measurements have been provided yet.
+        NotImplementedError: If the provided explainer does not support
+            the campaign surrogate.
+        TypeError: If the provided explainer does not support the campaign surrogate.
     """
     if campaign.measurements.empty:
         raise ValueError("No measurements have been provided yet.")
@@ -54,13 +57,27 @@ def explainer(
 
             return output.detach().numpy()
 
-    shap_explainer = explainer_class(model, data, **kwargs)
+    if (
+        campaign.searchspace.type != "CONTINUOUS"
+        and not computational_representation
+        and not explainer_class == shap.KernelExplainer
+    ):
+        raise NotImplementedError(
+            "Only KernelExplainer is supported for non-continous searchspaces."
+        )
+
+    try:
+        shap_explainer = explainer_class(model, data, **kwargs)
+    except shap.utils._exceptions.InvalidModelError:
+        raise TypeError(
+            "The selected explainer class does not support the campaign surrogate."
+        )
     return shap_explainer
 
 
 def explanation(
     campaign: Campaign,
-    data: np.ndarray = None,
+    data: pd.DataFrame | None = None,
     explainer_class: shap.Explainer = shap.KernelExplainer,
     computational_representation: bool = False,
     **kwargs,
@@ -82,8 +99,12 @@ def explanation(
         The Shapley values for the provided campaign.
 
     Raises:
+        ValueError: If the provided explainer does not support experimental
+            representation.
+        NotImplementedError: If the provided explainer does not support
+            the campaign surrogate.
         ValueError: If the provided data does not have the same amount of parameters
-            as previously provided to the explainer.
+            as the campaign.
     """
     is_shap_explainer = not explainer_class.__module__.startswith(
         "shap.explainers.other."
@@ -95,67 +116,82 @@ def explanation(
             "supported for non-Kernel SHAP explainer."
         )
 
-    explainer_obj = explainer(
-        campaign,
-        explainer_class=explainer_class,
-        computational_representation=computational_representation,
-        **kwargs,
-    )
+    try:
+        explainer_obj = explainer(
+            campaign,
+            explainer_class=explainer_class,
+            computational_representation=computational_representation,
+            **kwargs,
+        )
+    except NotImplementedError:
+        warnings.warn(
+            "The provided Explainer does not support experimental representation. "
+            "Switching to computational representation. "
+            "Otherwise consider using a different explainer (e.g. KernelExplainer)."
+        )
+        return explanation(
+            campaign,
+            data=data,
+            explainer_class=explainer_class,
+            computational_representation=True,
+            **kwargs,
+        )
 
     if data is None:
-        if isinstance(explainer_obj.data, np.ndarray):
-            data = explainer_obj.data
-        else:
-            data = explainer_obj.data.data
-    elif computational_representation:
-        data = campaign.searchspace.transform(data)
+        data = campaign.measurements[[p.name for p in campaign.parameters]].copy()
+    elif set(campaign.searchspace.parameter_names) != set(data.columns.values):
+        raise ValueError(
+            "The provided data does not have the same amount of parameters "
+            "as specified for the campaign."
+        )
+    if computational_representation:
+        data = campaign.searchspace.transform(pd.DataFrame(data))
+
+    """Get background data depending on the explainer."""
+    bg_data = getattr(explainer_obj, "data", getattr(explainer_obj, "masker", None))
+    bg_data = getattr(bg_data, "data", bg_data)
+
+    # Type checking for mypy
+    bg_data = bg_data if isinstance(bg_data, pd.DataFrame) else pd.DataFrame(bg_data)
+    assert isinstance(data, pd.DataFrame)
+
+    if not bg_data.shape[1] == data.shape[1]:
+        raise ValueError(
+            "The provided data does not have the same amount of "
+            "parameters as the shap explainer background."
+        )
 
     if not is_shap_explainer:
         """Return attributions for non-SHAP explainers."""
         if explainer_class.__module__.endswith("maple"):
             """Additional argument for maple to increase comparability to SHAP."""
-            attributions = explainer_obj.attributions(data, multiply_by_input=True)[0]
+            attributions = explainer_obj.attributions(
+                np.array(data), multiply_by_input=True
+            )[0]
         else:
-            attributions = explainer_obj.attributions(data)[0]
+            attributions = explainer_obj.attributions(np.array(data))[0]
         if computational_representation:
             feature_names = campaign.searchspace.comp_rep_columns
         else:
             feature_names = campaign.searchspace.parameter_names
         explanations = shap.Explanation(
             values=attributions,
-            base_values=data,
-            data=data,
+            base_values=np.array(data),
+            data=np.array(data),
         )
         explanations.feature_names = list(feature_names)
         return explanations
 
-    if data.shape[1] != explainer_obj.data.data.shape[1]:
-        raise ValueError(
-            "The provided data does not have the same amount "
-            "of parameters as the shap explainer background."
-        )
+    shap_explanations = explainer_obj(np.array(data))
+    if len(shap_explanations.shape) == 2:
+        return shap_explanations
     else:
         shap_explanations = explainer_obj(data)[:, :, 0]
 
     return shap_explanations
 
 
-def plot_beeswarm(explanation: shap.Explanation, **kwargs) -> None:
-    """Plot the Shapley values using a beeswarm plot."""
-    shap.plots.beeswarm(explanation, **kwargs)
-
-
-def plot_waterfall(explanation: shap.Explanation, **kwargs) -> None:
-    """Plot the Shapley values using a waterfall plot."""
-    shap.plots.waterfall(explanation, **kwargs)
-
-
-def plot_bar(explanation: shap.Explanation, **kwargs) -> None:
-    """Plot the Shapley values using a bar plot."""
-    shap.plots.bar(explanation, **kwargs)
-
-
-def plot_scatter(explanation: shap.Explanation | memoryview, **kwargs) -> None:
+def plot_shap_scatter(explanation: shap.Explanation, **kwargs) -> None:
     """Plot the Shapley values using a scatter plot while leaving out string values.
 
     Args:
@@ -176,7 +212,7 @@ def plot_scatter(explanation: shap.Explanation | memoryview, **kwargs) -> None:
     def is_not_numeric_column(col):
         return np.array([not isinstance(v, numbers.Number) for v in col]).any()
 
-    if data.ndim == 1:
+    if np.ndim(data) == 1:
         if is_not_numeric_column(data):
             warnings.warn(
                 "Cannot plot scatter plot for the provided "
@@ -185,6 +221,9 @@ def plot_scatter(explanation: shap.Explanation | memoryview, **kwargs) -> None:
         else:
             shap.plots.scatter(explanation, **kwargs)
     else:
+        # Type checking for mypy
+        assert isinstance(data, np.ndarray)
+
         number_enum = [i for i, x in enumerate(data[1]) if not isinstance(x, str)]
         if len(number_enum) < len(explanation.feature_names):
             warnings.warn(